# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для керування новинами з авторизацією через Sanctum",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/public/news": {
            "get": {
                "tags": [
                    "Public News"
                ],
                "description": "Отримати список публічних новин без авторизації",
                "operationId": "27324ad3e9adf2cc9203797e31c74f31",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Пошук за назвою або описом",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "технологія"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Фільтр за ID автора",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Дата початку (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Дата кінця (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-12-31"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список публічних новин",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NewsResource"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Невалідні параметри запиту",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Помилка валідації"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Помилка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Помилка при отриманні новин"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/public/news/{id}": {
            "get": {
                "tags": [
                    "Public News"
                ],
                "description": "Отримати деталі однієї публічної новини",
                "operationId": "22142c71fa39564f383850f69fd82df3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новини",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деталі новини",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Новину не знайдено або вона не опублікована",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Новину не знайдено або вона не опублікована"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Помилка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Помилка при отриманні новини"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContentBlockResource": {
                "title": "Content Block Resource",
                "description": "Content block resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "text",
                            "image",
                            "text_image_left",
                            "text_image_right"
                        ],
                        "example": "text"
                    },
                    "text_content": {
                        "type": "string",
                        "example": "Текст контентного блоку",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/content_blocks/image.png",
                        "nullable": true
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 12:00:00"
                    }
                },
                "type": "object"
            },
            "NewsResource": {
                "title": "News Resource",
                "description": "News resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Нова технологія в Laravel"
                    },
                    "image": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/news/image.png",
                        "nullable": true
                    },
                    "short_description": {
                        "type": "string",
                        "example": "Короткий опис новини про нові можливості"
                    },
                    "is_published": {
                        "type": "boolean",
                        "example": true
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 12:00:00",
                        "nullable": true
                    },
                    "author": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Іван Петренко"
                            },
                            "avatar": {
                                "type": "string",
                                "example": "http://localhost:8000/storage/avatars/avatar.png",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "content_blocks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContentBlockResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введіть токен отриманий після логіну/реєстрації",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}